@@ -1,8 +1,8 @@
-from collections.abc import Generator
+from collections.abc import Callable, Generator
 from datetime import date
 from functools import partial
 from pathlib import Path
-from typing import Any, Callable, Optional
+from typing import Any
 
 from qualibrate_app.api.core.domain.local_storage._id_to_local_path import (
     IdToLocalPath,
@@ -12,10 +12,10 @@ from qualibrate_app.api.core.utils.path.node import NodePath
 from qualibrate_app.api.core.utils.path.node_date import NodesDatePath
 
 
-def find_latest_node(base_path: Path) -> Optional[NodePath]:
+def find_latest_node(base_path: Path) -> NodePath | None:
     return max(
         map(NodePath, base_path.glob("*/#*")),
-        key=lambda p: p.id or -1,  # type: ignore[union-attr]
+        key=lambda p: p.id or -1,
         default=None,
     )
 
@@ -27,8 +27,8 @@ def find_latest_node_id(base_path: Path) -> IdType:
 
 def _validate_date_range(
     date_path: NodesDatePath,
-    min_date: Optional[date] = None,
-    max_date: Optional[date] = None,
+    min_date: date | None = None,
+    max_date: date | None = None,
 ) -> bool:
     try:
         dt = date_path.date
@@ -57,7 +57,7 @@ def find_n_latest_nodes_ids(
     page: int,
     per_page: int,
     project_name: str,
-    max_node_id: Optional[int] = None,
+    max_node_id: int | None = None,
 ) -> Generator[IdType, None, None]:
     """
     Generator of n latest nodes ids
