@@ -1,5 +1,3 @@
-from typing import Optional
-
 from qualibrate_app.api.core.models.base import ModelWithId
 from qualibrate_app.api.core.models.snapshot import (
     SimplifiedSnapshotWithMetadata,
@@ -9,4 +7,4 @@ from qualibrate_app.api.core.models.storage import Storage
 
 class Node(ModelWithId):
     snapshot: SimplifiedSnapshotWithMetadata
-    storage: Optional[Storage]
+    storage: Storage | None
