@@ -1,10 +1,10 @@
 from collections.abc import Mapping
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any
 
 from pydantic import BaseModel
 
 
 class Storage(BaseModel):
-    path: Optional[Path] = None
-    data: Optional[Mapping[str, Any]] = None
+    path: Path | None = None
+    data: Mapping[str, Any] | None = None
