@@ -1,14 +1,14 @@
-from collections.abc import Mapping, Sequence
+from collections.abc import Callable, Mapping, Sequence
 from itertools import chain
-from typing import Any, Callable, Optional, Union, cast
+from typing import Any, cast
 
 from qualibrate_app.api.core.types import DocumentSequenceType
 
 
 def _get_subpath_value_wildcard(
-    obj: Union[Mapping[str, Any], Sequence[Any]],
-    target_path: Sequence[Union[str, int]],
-    current_path: list[Union[str, int]],
+    obj: Mapping[str, Any] | Sequence[Any],
+    target_path: Sequence[str | int],
+    current_path: list[str | int],
 ) -> DocumentSequenceType:
     if len(target_path) == 1:
         if isinstance(obj, Sequence):
@@ -27,9 +27,7 @@ def _get_subpath_value_wildcard(
     if not isinstance(obj, (Mapping, Sequence)):
         return []
     iter_function = cast(
-        Callable[
-            ..., Union[Sequence[tuple[str, Any]], Sequence[tuple[int, Any]]]
-        ],
+        Callable[..., Sequence[tuple[str, Any]] | Sequence[tuple[int, Any]]],
         enumerate if isinstance(obj, Sequence) else dict.items,
     )
     return list(
@@ -40,7 +38,7 @@ def _get_subpath_value_wildcard(
     )
 
 
-def _check_key_valid(obj: Any, key: Union[str, int]) -> bool:
+def _check_key_valid(obj: Any, key: str | int) -> bool:
     if isinstance(obj, Sequence):
         return isinstance(key, int) and 0 <= key < len(obj)
     if isinstance(obj, Mapping):
@@ -50,8 +48,8 @@ def _check_key_valid(obj: Any, key: Union[str, int]) -> bool:
 
 def get_subpath_value_mapping(
     obj: Mapping[str, Any],
-    target_path: Sequence[Union[str, int]],
-    current_path: list[Union[str, int]],
+    target_path: Sequence[str | int],
+    current_path: list[str | int],
     key: str,
 ) -> DocumentSequenceType:
     if len(target_path) == 1:
@@ -64,8 +62,8 @@ def get_subpath_value_mapping(
 
 def get_subpath_value_sequence(
     obj: Sequence[Any],
-    target_path: Sequence[Union[str, int]],
-    current_path: list[Union[str, int]],
+    target_path: Sequence[str | int],
+    current_path: list[str | int],
     key: int,
 ) -> DocumentSequenceType:
     if len(target_path) == 1:
@@ -77,9 +75,9 @@ def get_subpath_value_sequence(
 
 
 def get_subpath_value(
-    obj: Union[Mapping[str, Any], Sequence[Any]],
-    target_path: Sequence[Union[str, int]],
-    current_path: Optional[list[Union[str, int]]] = None,
+    obj: Mapping[str, Any] | Sequence[Any],
+    target_path: Sequence[str | int],
+    current_path: list[str | int] | None = None,
 ) -> DocumentSequenceType:
     if current_path is None:
         current_path = []
@@ -101,10 +99,10 @@ def get_subpath_value(
 
 
 def get_subpath_value_on_any_depth(
-    obj: Union[Mapping[str, Any], Sequence[Any]],
+    obj: Mapping[str, Any] | Sequence[Any],
     key: str,
-    current_path: Optional[list[Union[str, int]]] = None,
-    paths: Optional[list[dict[str, Any]]] = None,
+    current_path: list[str | int] | None = None,
+    paths: list[dict[str, Any]] | None = None,
 ) -> list[dict[str, Any]]:
     if current_path is None:
         current_path = []
