@@ -1,5 +1,5 @@
 from collections.abc import Sequence
-from typing import Annotated, Union
+from typing import Annotated
 
 from fastapi import Query
 
@@ -28,7 +28,7 @@ def check_path(value: str) -> str:
 
 def get_search_path(
     data_path: Annotated[str, Query(description="Path to search")],
-) -> Sequence[Union[str, int]]:
+) -> Sequence[str | int]:
     """Generate list of search subpaths from search path.
 
     Raises:
