@@ -554,7 +554,9 @@ def test_branch_nodes_history_default_args(
         "total_pages": 1,
         "items": [
             {"id": snapshot["id"], "snapshot": snapshot, "storage": dfs}
-            for snapshot, dfs in zip(snapshots_history, dfss_history)
+            for snapshot, dfs in zip(
+                snapshots_history, dfss_history, strict=False
+            )
         ],
     }
 
@@ -574,7 +576,7 @@ def test_branch_nodes_history_reverse(
         "items": [
             {"id": snapshot["id"], "snapshot": snapshot, "storage": dfs}
             for snapshot, dfs in zip(
-                snapshots_history[::-1], dfss_history[::-1]
+                snapshots_history[::-1], dfss_history[::-1], strict=False
             )
         ],
     }
@@ -596,7 +598,7 @@ def test_branch_nodes_history_reverse_paged(
         "items": [
             {"id": snapshot["id"], "snapshot": snapshot, "storage": dfs}
             for snapshot, dfs in zip(
-                snapshots_history[3:1:-1], dfss_history[3:1:-1]
+                snapshots_history[3:1:-1], dfss_history[3:1:-1], strict=False
             )
         ],
     }
@@ -618,7 +620,7 @@ def test_branch_nodes_history_global_reverse(
         "items": [
             {"id": snapshot["id"], "snapshot": snapshot, "storage": dfs}
             for snapshot, dfs in zip(
-                snapshots_history[::-1], dfss_history[::-1]
+                snapshots_history[::-1], dfss_history[::-1], strict=False
             )
         ],
     }
@@ -641,7 +643,7 @@ def test_branch_nodes_history_global_reverse_paged(
         "items": [
             {"id": snapshot["id"], "snapshot": snapshot, "storage": dfs}
             for snapshot, dfs in zip(
-                snapshots_history[7:5:-1], dfss_history[7:5:-1]
+                snapshots_history[7:5:-1], dfss_history[7:5:-1], strict=False
             )
         ],
     }
@@ -685,7 +687,7 @@ def test_branch_nodes_history_paged(
             {"id": snapshot["id"], "snapshot": snapshot, "storage": dfs}
             for _, (snapshot, dfs) in filter(
                 lambda d: d[0] in requested_ids,
-                enumerate(zip(snapshots_history, dfss_history)),
+                enumerate(zip(snapshots_history, dfss_history, strict=False)),
             )
         ],
     }
