@@ -65,13 +65,13 @@ def test_parse_float(data, expected):
 @pytest.mark.parametrize(
     "data, expected",
     (
-        (1.0, "1.0"),
-        (1, "1"),
+        (1.0, 1.0),
+        (1, 1),
         ("2", "2"),
         ("-3", "-3"),
         ("true", "true"),
         ("random", "random"),
-        (b"random", "b'random'"),
+        (b"random", b"random"),
     ),
 )
 def test_parse_str(data, expected):
