@@ -1,6 +1,5 @@
 import React from "react";
 import { JSONEditor } from "../../../Data/components/JSONEditor";
-import { useNodesContext } from "../../context/NodesContext";
 import styles from "./Results.module.scss";
 import { ModuleKey } from "../../../../routing/ModulesRegistry";
 import ErrorIcon from "../../../../ui-lib/Icons/ErrorIcon";
@@ -35,19 +34,13 @@ export const ResultsError: React.FC<{ style?: React.CSSProperties; errorObject: 
 
 export const Results: React.FC<{
   title?: string;
-  jsonObject?: unknown;
+  jsonObject: object;
   showSearch?: boolean;
   toggleSwitch?: boolean;
   pageName?: ModuleKey;
   errorObject: ErrorObject | null | undefined;
   style?: React.CSSProperties;
 }> = ({ title, jsonObject, showSearch = true, toggleSwitch = false, pageName, style, errorObject }) => {
-  let jsonData = jsonObject;
-  if (!jsonObject) {
-    const { results } = useNodesContext();
-    jsonData = results;
-  }
-
   if (errorObject) {
     return <ResultsError style={style} errorObject={errorObject} />;
   }
@@ -56,7 +49,7 @@ export const Results: React.FC<{
     <div className={styles.wrapper} style={style} data-testid="results-wrapper">
       <JSONEditor
         title={title ?? "Results"}
-        jsonDataProp={jsonData ?? {}}
+        jsonDataProp={jsonObject}
         height={"100%"}
         showSearch={showSearch}
         toggleSwitch={toggleSwitch}
