@@ -5,12 +5,11 @@ import { useFlexLayoutContext } from "../../routing/flexLayout/FlexLayoutContext
 import modulesMap, { PROJECT_KEY } from "../../routing/ModulesRegistry";
 import PageName from "../../common/ui-components/common/Page/PageName";
 import TitleBarGraphCard from "./TitleBarGraphCard/TitleBarGraphCard";
-import ProjectTitleBar from "./ProjectTitleBar";
 
 const TopBar: React.FC = () => {
   const { activeTab } = useFlexLayoutContext();
 
-  return activeTab === PROJECT_KEY ? <ProjectTitleBar /> : <TitleBarGraphCard />;
+  return activeTab === PROJECT_KEY ? null : <TitleBarGraphCard />;
 };
 
 const TitleBarMenu: React.FC = () => {
